:recycle: Renamed components folders
