+++ src/lib/FloatingButton/index.js
